# Introduce cache to speed things up and cleanup method