Whatsup folks! A colleague of mine added puppetdb into our tests and I am extremely happy to see more test coverage thanks to this wonderful effort. But our tests are now a minute slower. We can do better, come on! :-)

This patch helps a bit with two problems - CPU and memory.

First, from what I understand every filter string should generate the very same result. So why not to cache both input (filter string) and output? I use thread-local variable for that. It memorizes the last filter seen and if there is a match, it quickly returns the filtered data. This will be also thread-safe.

Now this will only help in cases when facts are requested with same filter, however it looks like that is exactly what we do in our test suite. We construct facterdb and then get facts for several combinations of OSes testing one fact after another. This patch shaves off a lot of time in our case. And if you think our project might be some kind of special case, well, I measured tests of the facterdb itself.

Without this patch:

```
real	0m12,418s
user	0m11,805s
sys	0m0,508s
```

With this patch:

```
real	0m6,395s
user	0m5,799s
sys	0m0,514s
```

That's a reasonable 50% speedup if I do my math correctly. Not bad for Friday, not bad. I am not stopping there tho.

Second, once facts are loaded into memory and stored in the `@database` field, there is no way to reclaim the memory back. The files currently weights about 15MB and it will probably only grow. So I added a new method called `cleanup` which allows Ruby VM to reclaim this memory back. Not problem for this project, but our project has about 4000 other tests to execute after its done with facterdb. Typically users will call this method in global teardown or after(:suite) or similar.

I also have a suggestion which I am not going to work on: I understand that the database is populated into one huge JSON string during runtime. Why this isn't part of the build process? That huge JSON could be distributed in the gem and built via a Rake task so there is no need to load it from individual files. Just a tought.

Cheers!