@ekohl 2021-06-18
This may make https://github.com/voxpupuli/rspec-puppet-facts/commit/43bb17e9f894421daef73e833ca3ea0a506f8d9c less needed. One thing to note is that there I chose to use marshal as a way to deep clone in order to prevent in place modifications would corrupt the cache.