@ekohl 2021-06-18
> I also have a suggestion which I am not going to work on: I understand that the database is populated into one huge JSON string during runtime. Why this isn't part of the build process? That huge JSON could be distributed in the gem and built via a Rake task so there is no need to load it from individual files. Just a tought.

That's an interesting thought. Looks like it'll require some fiddling in the gemspec (now it uses git ls-files) and the release step.

This gem was just migrated to Voxpupuli last week and we haven't had a chance to convert from Travis to Github Actions, but the release actions we use should make it quite easy to add another file. I'll look into this.