@lzap 2021-06-18
> This may make [voxpupuli/rspec-puppet-facts@43bb17e](https://github.com/voxpupuli/rspec-puppet-facts/commit/43bb17e9f894421daef73e833ca3ea0a506f8d9c) less needed

Probably, on the other hand the `cleanup` method is something that should be probably added into there so users do not need to care.

> I'll look into this.

It might not speedup things much, I think the biggest offender is the JGrep but I might be wrong.

If there is a chance to see a release after this is merged, we would be able to rebase our patch and use the `cleanup` method from the day one. That would be cool.