@bastelfreak 2021-06-18
@lzap thanks for the PR. Can you please rebase? That should enable CI again